Vendor upstream container build
Problem:
The runner version isn't easy to update using an overlay dockerfile.

Solution:
Vendor the entire upstream container image build process.

Result:
This isn't ideal, or recommended from the detailed documentation in the
actions-runner-controller, but it works and helps explore how image
building works and can be customized.

The intent is to switch back to building from the upstream image once
it's updated more quickly.